refact clickhouse sink
