+++ src/sinks/clickhouse/mod.rs
@@ -1,0 +1,5 @@
+mod config;
+mod http_sink;
+#[cfg(all(test, feature = "clickhouse-integration-tests"))]
+mod integration_tests;
+pub use self::config::ClickhouseConfig;
